third_party/WebKit/LayoutTests/platform/win/resize-observer/eventloop-expected.txt

CONSOLE ERROR: ResizeObserver loop limit exceeded
CONSOLE ERROR: ResizeObserver loop limit exceeded
This is a testharness.js-based test.
PASS guard 
PASS test0: multiple notifications inside same event loop 
PASS test1: depths of shadow roots 
PASS test2: move target in dom while inside event loop 
Harness: the test ran to completion.
